+++ src/__tests__/completion/language.test.ts
@@ -1,4 +1,4 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import { CancellationToken, Disposable } from 'vscode-languageserver-protocol'
 import { CompletionItem, CompletionItemKind, CompletionList, InsertReplaceEdit, InsertTextFormat, InsertTextMode, Position, Range, TextEdit } from 'vscode-languageserver-types'
 import commandManager from '../../commands'
@@ -614,6 +614,14 @@ describe('language source', () => {
       })
       await helper.confirmCompletion(0)
       await helper.waitFor('getline', ['.'], 'foo()')
+    })
+
+    it('should use textEditText when exists with default range', async () => {
+      await start({ label: 'foo', insertText: 'bar', textEditText: 'foofoo' }, {
+        editRange: Range.create(0, 0, 0, 0)
+      })
+      await helper.confirmCompletion(0)
+      await helper.waitFor('getline', ['.'], 'foofoo')
     })
   })