+++ src/__tests__/handler/locations.test.ts
@@ -1,4 +1,4 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import { Disposable, LocationLink, Location, Range, Position, CancellationTokenSource, CancellationToken } from 'vscode-languageserver-protocol'
 import LocationHandler from '../../handler/locations'
 import languages from '../../languages'
@@ -278,10 +278,10 @@ describe('locations', () => {
       expect(res).toEqual([
         {
           name: 'foo',
-          cmd: 'silent keepjumps call cursor(3, 1)',
+          cmd: 'silent keepjumps call coc#cursor#move_to(2, 0)',
           filename: 'test://bar'
         },
-        { name: 'foo', cmd: 'silent keepjumps call cursor(2, 1)', filename: '/foo' }
+        { name: 'foo', cmd: 'silent keepjumps call coc#cursor#move_to(1, 0)', filename: '/foo' }
       ])
     })
   })