+++ src/__tests__/modules/extensionModules.test.ts
@@ -1,10 +1,11 @@
 process.env.COC_NO_PLUGINS = '1'
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import fs from 'fs'
 import os from 'os'
 import path from 'path'
 import { v4 as uuid } from 'uuid'
 import { Disposable } from 'vscode-languageserver-protocol'
+import { URI } from 'vscode-uri'
 import events from '../../events'
 import { checkExtensionRoot, ExtensionStat, getExtensionName, getJsFiles, loadExtensionJson, loadGlobalJsonAsync, toInterval, validExtensionFolder } from '../../extension/stat'
 import { InstallBuffer, InstallChannel } from '../../extension/ui'
@@ -215,6 +216,15 @@ describe('ExtensionStat', () => {
     expect(stat.getFolder('unknown')).toBeUndefined()
   })
 
+  it('should addNoPromptFolder', async () => {
+    let [state, filepath] = create()
+    let uri = URI.file(path.dirname(filepath)).toString()
+    expect(state.shouldPrompt(uri)).toBe(true)
+    state.addNoPromptFolder(uri)
+    state.addNoPromptFolder(uri)
+    expect(state.shouldPrompt(uri)).toBe(false)
+  })
+
   it('should iterate activated extensions', () => {
     let folder = createFolder()
     writeJson(path.join(folder, 'package.json'), {
@@ -387,7 +397,7 @@ describe('InstallBuffer', () => {
 
   it('should sync by not split', async () => {
     global.__TEST__ = false
-    let buf = new InstallBuffer(false)
+    let buf = new InstallBuffer({ isUpdate: false, updateUIInTab: false })
     disposables.push(buf)
     events.requesting = true
     await buf.start(['a', 'b', 'c'])
@@ -397,7 +407,7 @@ describe('InstallBuffer', () => {
   })
 
   it('should draw buffer with stats', async () => {
-    let buf = new InstallBuffer(true)
+    let buf = new InstallBuffer({ isUpdate: true, updateUIInTab: true })
     disposables.push(buf)
     buf.draw()
     await buf.start(['a', 'b', 'c', 'd'])
@@ -418,7 +428,7 @@ describe('InstallBuffer', () => {
   })
 
   it('should stop when all items finished', async () => {
-    let buf = new InstallBuffer(false)
+    let buf = new InstallBuffer({ isUpdate: false })
     disposables.push(buf)
     await buf.start(['a', 'b'])
     buf.startProgress('a')
@@ -433,7 +443,7 @@ describe('InstallBuffer', () => {
 
   it('should show messages and dispose', async () => {
     events.requesting = true
-    let buf = new InstallBuffer(true)
+    let buf = new InstallBuffer({ isUpdate: true })
     disposables.push(buf)
     await buf.start(['a', 'b'])
     buf.startProgress('a')
@@ -456,7 +466,7 @@ describe('InstallBuffer', () => {
 describe('InstallChannel', () => {
   it('should create install InstallChannel', async () => {
     let outputChannel = window.createOutputChannel('test')
-    let channel = new InstallChannel(true, outputChannel)
+    let channel = new InstallChannel({ isUpdate: true }, outputChannel)
     channel.start(['a', 'b'])
     channel.startProgress('a')
     channel.addMessage('a', 'msg', true)
@@ -467,7 +477,7 @@ describe('InstallChannel', () => {
 
   it('should create update InstallChannel', async () => {
     let outputChannel = window.createOutputChannel('test')
-    let channel = new InstallChannel(false, outputChannel)
+    let channel = new InstallChannel({ isUpdate: false }, outputChannel)
     channel.start(['a', 'b'])
     channel.startProgress('a')
     channel.finishProgress('a', true)