+++ src/__tests__/snippets/session.test.ts
@@ -1,4 +1,4 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import path from 'path'
 import { Position, Range, TextEdit } from 'vscode-languageserver-protocol'
 import { UltiSnippetContext } from '../../snippets/eval'
@@ -132,6 +132,8 @@ describe('SnippetSession', () => {
       expect(lines).toEqual([
         'fun', '  abc', '    def', 'end'
       ])
+      let val = await nvim.getVar('coc_selected_text')
+      expect(val).toBe(null)
     })
 
     it('should resolve VISUAL', async () => {
@@ -361,8 +363,9 @@ describe('SnippetSession', () => {
       await session.start('${1:foo} bar$0', defaultRange)
       await nvim.input('<backspace>')
       await session.forceSynchronize()
-      let col = await nvim.call('col', ['.'])
-      expect(col).toBe(5)
+      await helper.waitValue(async () => {
+        return await nvim.call('col', ['.'])
+      }, 5)
     })
 
     it('should prefer range contains current cursor', async () => {
@@ -443,7 +446,7 @@ describe('SnippetSession', () => {
       let res = await session.start('a${1:a}b', defaultRange)
       expect(res).toBe(true)
       await buf.append(['foo', 'bar'])
-      await nvim.call('cursor', [2, 1])
+      await nvim.call('cursor', [2, 2])
       await session.checkPosition()
       expect(session.isActive).toBe(false)
     })
@@ -531,6 +534,22 @@ describe('SnippetSession', () => {
       await session.nextPlaceholder()
       let col = await nvim.call('col', '.')
       expect(col).toBe(5)
+    })
+
+    it('should remove white space on jump', async () => {
+      let session = await createSession()
+      let opts = {
+        removeWhiteSpace: true,
+        ...defaultContext
+      }
+      let res = await session.start('foo  $1\n${2:bar} $0', defaultRange, true, opts)
+      expect(res).toBe(true)
+      let line = await nvim.line
+      expect(line).toBe('foo  ')
+      await session.nextPlaceholder()
+      expect(session.isActive).toBe(true)
+      let lines = await session.document.buffer.lines
+      expect(lines[0]).toBe('foo')
     })
   })